client-core: teach the client core about V3 handles.
This is a first check-in to my V3 client-core feature branch. Many of the
places that were changed have the old code commented out in-place so that
the new code can be more easily found and reviewed. After being tested
and fixed up wherever needed, it will be cleaned up and finally
commited to next0...

Basically, we used to have to turn 64 bit handles into 128 bit
khandles when preparing the downcall, and back from 128 to 64 when
receiving the upcall. Now that handles are all 128 bit uuids,
they "already are" khandles. Khandles still have to be converted
to a 64 bit inode number in the kernel module, this remains unchanged,
and hopefully the upstream kernel module will still work <g> ...

SVN revision 13095